Post code prevedeny na int z decimalu

+++ Module.Order/Dal/Entities/Address.cs
@@ -9,7 +9,7 @@ namespace Module.Order.Dal.Entities
     public class Address
     {
         public Address() { }
-        public Address (string street, string city, int house_number, decimal post_code, int country)
+        public Address (string street, string city, int house_number, int post_code, int country)
         {
             this.street = street;
             this.city = city;
@@ -39,7 +39,7 @@ namespace Module.Order.Dal.Entities
 
         public string city { get; set; }
         public int house_number { get; set; }
-        public decimal post_code { get; set; }
+        public int post_code { get; set; }
         
         [ForeignKey("code")]
         public int country { get; set; }
